+++ kodi_src/xbmc/xbmc/filesystem/IFile.h
@@ -85,7 +85,7 @@ public:
    *         buffer, zero if no bytes are available to read (end of file was reached)
    *         or undetectable error occur, -1 in case of any explicit error
    */
-  virtual ssize_t Read(void* bufPtr, size_t bufSize) = 0;
+  virtual int64_t Read(void* bufPtr, size_t bufSize) = 0;
   /**
    * Attempt to write bufSize bytes from buffer bufPtr into currently opened file.
    * @param bufPtr  pointer to buffer
@@ -94,7 +94,7 @@ public:
    *         zero if no bytes were written and no detectable error occur,
    *         -1 in case of any explicit error
    */
-  virtual ssize_t Write(const void* bufPtr, size_t bufSize) { return -1;}
+  virtual int64_t Write(const void* bufPtr, size_t bufSize) { return -1;}
   virtual bool ReadString(char *szLine, int iLineLength);
   virtual int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET) = 0;
   virtual void Close() = 0;
