+++ kodi_src/xbmc/xbmc/filesystem/ShoutcastFile.h
@@ -32,7 +32,7 @@ public:
   bool Open(const GoTvUrl& url) override;
   bool Exists(const GoTvUrl& url) override { return true;};
   int Stat(const GoTvUrl& url, struct __stat64* buffer) override { errno = ENOENT; return -1; };
-  ssize_t Read(void* lpBuf, size_t uiBufSize) override;
+  int64_t Read(void* lpBuf, size_t uiBufSize) override;
   int64_t Seek(int64_t iFilePosition, int iWhence = SEEK_SET) override;
   void Close() override;
   int IoControl(EIoControl request, void* param) override;
